+++ src/base/relation.js
@@ -1,7 +1,7 @@
 // Base Relation
 // ---------------
 
-import _, { assign } from 'lodash';
+import { assign, result } from 'lodash';
 import CollectionBase from './collection';
 import extend from '../extend';
 
@@ -11,8 +11,8 @@ export default class RelationBase {
 
   constructor(type, Target, options) {
     if (Target != null) {
-      this.targetTableName   = _.result(Target.prototype, 'tableName');
-      this.targetIdAttribute = _.result(Target.prototype, 'idAttribute');
+      this.targetTableName   = result(Target.prototype, 'tableName');
+      this.targetIdAttribute = result(Target.prototype, 'idAttribute');
     }
     assign(this, { type, target: Target }, options);
   }